Se dividio la plantilla de usuarios y usuarios en espera
